Handle the start and rows parameters correctly
Chef Zero would, on a search request using the "start" or "rows"
parameters, always return all rows.

This change makes it so only the correct number of rows are returned or
the results start at the correct index.

This change will make the tests added in oc-chef-pedant in
chef/chef-server#1028 pass.

Also update the Gemfile to use HTTPS URLs for Git sources, remove unused
comments, and use the Chef Gem from RubyGems as the previous comment
suggested.

Signed-off-by: Nathan L Smith <smith@chef.io>